+++ RELEASE_NOTES.md
@@ -1,6 +1,8 @@
 ## 0.37.0 (2024-04-27)
 - **BREAKER** Functionally everything about this version
 - Vastly fewer allocations due to data restructuring
+- `server.Reqlen` renamed to `server.Xferlim`
+- `Xferlim` added to `client.Config`
 - `p.Perr` renamed to `p.Status`; `p.Errs` renamed to `p.Stats`
 - Status remappings:
   - 196 is now 498